project_ML/requirements.txt
keras==2.10.0
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
pandas_datareader==0.10.0
scikit_learn==1.1.3
imutils==0.5.4
opencv-python
tensorflow_datasets==3.2.1



